Refactor code and improve UI UX performance
This commit involves various code refactors spanning over several files with changes primarily directed towards improving the UI/UX performance of the application. The changes include the deletion of the about.tsx page to enhance focus on the core components of the project. Other updates include correcting typos and syntax errors in various tsx component files.

The animatedText.tsx file has been modified to include missing dependencies in the useEffect hook to guard against unnecessary re-renders. The naming convention for props in a few components was not consistent and have been updated to semantically correct versions (for example, the 'offset' prop was changed to 'offSetY') which renders the code easier to read and maintain.

The visual appeal of the project pages has been improved by reordering elements of the twoColumnLayout.tsx file to render a more intuitive and user-friendly layout. Additional enhancements involve making the animated gradient background more engaging through the application of a more dynamic linear gradient.

Overall, these changes aimed at improving the readability of the code, enhancing the scalability of the components and uplifting the user interface of the projects.